CCA: Move CCA MK change code into separate source file
No logic change, just moving existing code into a new source file
cca_mkchange.c. This requires that certain definitions are moved into
cca_stdll.h, and some functions can no longer be static.
