+++ usr/lib/cca_stdll/cca_stdll.mk
@@ -45,7 +45,8 @@ opencryptoki_stdll_libpkcs11_cca_la_SOURCES = usr/lib/common/asn1.c	\
 	usr/lib/config/cfgparse.y usr/lib/config/cfglex.l		\
 	usr/lib/common/mech_openssl.c usr/lib/common/pqc_supported.c	\
 	usr/lib/hsm_mk_change/hsm_mk_change.c				\
-	usr/lib/common/btree.c usr/lib/common/sess_mgr.c
+	usr/lib/common/btree.c usr/lib/common/sess_mgr.c		\
+	usr/lib/cca_stdll/cca_mkchange.c
 
 if !NO_PKEY
 opencryptoki_stdll_libpkcs11_cca_la_SOURCES +=				\
